GPSMap/GPSMap_getMapBinary/requirements.txt
google-cloud-error-reporting==0.30.0
opencv-python==4.0.0.21
Pillow
googlemaps
